+++ hpOneView/resources/resource.py
@@ -375,6 +375,28 @@ class ResourceHelper(object):
 
         return self.do_requests_to_getall(uri, count)
 
+    def delete_all(self, filter, force=False, timeout=-1):
+        """
+        Deletes all resources from the appliance that match the provided filter.
+
+        Args:
+            filter:
+                A general filter/query string to narrow the list of items deleted.
+            force:
+                If set to true, the operation completes despite any problems with network connectivity or errors
+                on the resource itself. The default is false.
+            timeout:
+                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
+                in OneView; it just stops waiting for its completion.
+
+        Returns:
+            bool: Indicates if the resources were successfully deleted.
+        """
+        uri = "{}?filter={}&force={}".format(self._base_uri, quote(filter), force)
+        logger.debug("Delete all resources (uri = %s)" % uri)
+
+        return self.delete(uri)
+
     def create(self, data=None, uri=None, timeout=-1, custom_headers=None, force=False):
         """Makes a POST request to create a resource when a request body is required.
 
@@ -561,6 +583,13 @@ class ResourceHelper(object):
         uri = "{0}{1}start={2}&count={3}{4}{5}{6}{7}{8}{9}".format(path, symbol, start, count, filter, query, sort,
                                                                    view, fields, scope_uris)
         return uri
+
+    def build_uri_with_query_string(self, kwargs, sufix_path='', uri=None):
+        if not uri:
+            uri = self._base_uri
+
+        query_string = '&'.join('{}={}'.format(key, kwargs[key]) for key in sorted(kwargs))
+        return uri + sufix_path + '?' + query_string
 
     def build_uri(self, id_or_uri):
         """Helps to build the URI from resource id and validate the URI.