+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Resource data will be available with the resource object. This enhancement helps
 - SAS logical interconnect group
 - Server hardware
 - Server hardware type
+- Server profile
 - Server profile template
 - Switch type
 - Uplink set